Improved connection handling for workers
This work is the result of an outage we had at DealNews. The code has
been in use for about a year and is quite stable now.

We suffered some unavailable workers as a result of high latency. The
connections to the up server timed out due to over anxious attempts to
find work that caused gearmand to slow down and time out, thus putting
it in the retry list. However, the socket was not closed. This would
lead to lots of sockets open that did not have an object in PHP user
space. Rather than set the sleep time on the socket read when
sleeping, instead, loop for the sleep time and use the retry time as
the socket timeout. This will allow reconnects to happen sooner and
will also make the worker more responsive to shut down requests.
Properly closed the connection instead of relying on the unset() to
magically do it from the Worker code.