Better connection closing #34
The order of operations in Connection::close() was causing a
blocking operation even when the socket had closed fine. The code
now checks if the connection was closed before it tries to read
any pending data that may still be on the connection. Additionally,
there is now a low timeout set before attempting the close to
avoid long blocking.

A new callback has been added to Worker for reporting various
status information. This makes it easier to debug what is happening
inside the Worker object when it appears to be hanging.

The code now calls the monitor callback more often to ensure that
it responds to any signal to shutdown more promptly. This may have
been part of the problem with the CLOSE_WAIT issues that have been
seen in the past. During testing, a SIG_KILL of 60 workers that
were in a loop with usleep() led to 60 CLOSE_WAIT connections on
the gearmand server they were connected to. Using sleep in Worker
and checking the monitor more often appears to have made that issue
either disappear or much, much less likely to happen.